refactor(web): client/lib の未使用・単一利用コードを整理
- getApiErrorMessage はプロダクションコードからの参照がなく、自身のテストだけが
  使用していたため error.ts ごと削除
- validation のバリデーション結果ヘルパーは authenticate 機能のみが利用していたため、
  client/lib から features/authenticate 配下へ移設し共通基盤の間接層を解消
